Corrigindo Migration

+++ db/migrate/20110427115028_add_price_to_media.rb
@@ -1,6 +1,6 @@
 class AddPriceToMedia < ActiveRecord::Migration
   def self.up
-    add_column :medias, :price, :double
+    add_column :medias, :price, :decimal
   end
 
   def self.down
